fix: reorganise pulse util functions and fix graph max value
