+++ resources/js/components/dashboard/DashboardAnalytics.vue
@@ -4,7 +4,7 @@ import type { PulseResponse } from '@/types/types';
 import { useGetPulse, useGetSiteAnalytics } from '@/service/queries';
 import { ref, useTemplateRef, watch } from 'vue';
 import { handleStartTask } from '@/service/taskService';
-import { periodForHumans } from '@/service/util';
+import { periodForHumans } from '@/service/pulseUtil';
 
 import PulseRequests from '@/components/pulseCards/PulseRequests.vue';
 import DashboardCard from '@/components/cards/DashboardCard.vue';
